+++ how-to/src/components/HomePage/Cards/GuideCard.js
@@ -3,6 +3,11 @@ import Buttons from "./Buttons";
 import styled from "styled-components";
 
 // Styles
+const SectionCardStyle = styled.section`
+  display: flex;
+  justify-content: space-evenly;
+  align-content: center;
+`;
 const Card = styled.form`
   /* Main Card Styles */
   background: ${props => props.theme.primaryColor};
@@ -10,6 +15,7 @@ const Card = styled.form`
   color: ${props => props.theme.fontColorLight};
   width: 400px;
   padding: 3%;
+  margin: 2%;
 
   /* Elements */
   h1 {
@@ -22,23 +28,25 @@ const Card = styled.form`
     margin: 2%;
     color: ${props => props.theme.fontColorLight};
   }
+
+  p {
+    margin: 5%;
+  }
 `;
 
-const GuideCard = ({ guide }) => {
+const GuideCard = ({ guides }) => {
   return (
-    <>
+    <SectionCardStyle>
       <Card>
         <div>
-          <h1>
-            {guide.title}
-
-            <div>
-              <Buttons />
-            </div>
-          </h1>
+          <h1>{guides.title}</h1>
+          <div>
+            <Buttons state={guides} />
+          </div>
+          <p>{guides.Steps}</p>
         </div>
       </Card>
-    </>
+    </SectionCardStyle>
   );
 };
