+++ how-to/src/components/HomePage/Cards/GuideList.js
@@ -1,0 +1,16 @@
+import React, { useState, useEffect } from "react";
+import SearchForm from "../SearchBar/SearchBar";
+import GuideCard from "./GuideCard";
+
+const FormList = ({ data }) => {
+  return (
+    <>
+      <SearchForm />
+      {data.map(guide => {
+        return <GuideCard guides={guide} />;
+      })}
+    </>
+  );
+};
+
+export default FormList;
